requirements.txt
cassandra-driver[eventlet]>=3.28.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.2
numpy>=1.23.0
google-generativeai>=0.3.0
streamlit>=1.27.0
eventlet>=0.33.3 